Update hackers.md
this should be more clear about id and hacked columns,
id doesn't actually ID to the specific hacker, it IDs for the event its self (1 person can have multiple events)
hacked just said hacked, not a very detailed description of what exactly that means. i think this is a bit more clear as to what exactly that means.